Rename db file
