+++ services/blockchain-app-registry/shared/database/schema/application_metadata.js
