+++ services/blockchain-app-registry/shared/database/schema/token_metadata.js
